pyproject.toml
[project]
name = "discord_fish_game"
version =  "0.1.0"
dependencies = [
  "py-cord >= 2.6.1",
  "setproctitle >= 1.3.7",
  "python-dotenv >= 1.2.1",
  "psycopg >= 3.3.2",
  "bidict >= 0.1.5",
  "scipy >= 1.16.3"
]

